feat(data): remove need of fiches-sp.json
