+++ packages/code-du-travail-data/dataset/fiches_service_public/package.json
@@ -15,10 +15,6 @@
     "unist-util-find": "^1.0.1"
   },
   "scripts": {
-    "backup": "cp fiches-sp.json fiches-sp.old.json ",
-    "diff": "node diff.js",
-    "updateTheme": "node updateTheme.js > out.json && mv out.json fiches-sp.json  && yarn report",
-    "start": "yarn backup && node index.js > fiches-sp.json",
-    "report": "node diff.js && node missingTheme.js"
+    "report": "node missingTheme.js"
   }
 }
